+++ mocks/include/mocks/BaseApplication.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "common/Args.hpp"
 #include "mocks/DisabledStreamerMode.hpp"
 #include "mocks/EmptyApplication.hpp"
 #include "providers/bttv/BttvLiveUpdates.hpp"
@@ -16,13 +17,13 @@ class BaseApplication : public EmptyApplication
 {
 public:
     BaseApplication()
-        : settings(this->settingsDir.filePath("settings.json"))
+        : settings(this->args, this->settingsDir.path())
     {
     }
 
     explicit BaseApplication(const QString &settingsData)
         : EmptyApplication(settingsData)
-        , settings(this->settingsDir.filePath("settings.json"))
+        , settings(this->args, this->settingsDir.path())
     {
     }
 
@@ -41,6 +42,7 @@ public:
         return nullptr;
     }
 
+    Args args;
     Settings settings;
     DisabledStreamerMode streamerMode;
 };